refactor(nika-providers): the probe layer joins its owning crate — the cap forces the shape
The 15k wall bit on the PR's merge-ref (my context verb + the sister's
trace-retention arc summed past it): provider environment probing —
ProviderProbe · PingState · env_present · ping_addr · spawn_ping ·
collect_local_pings · collect_provider_probes — moves to
nika_providers::probe, the crate that owns env_candidates() and
effective_base_url(). The CLI keeps the render voices and the
client/media/pricing probes; the provider FACTS live with the
providers. Rode along: the mirror family dedups its workspace walk +
environment JSON fragment into the shared probe layer (welcome ·
context · one truth).

Baselines: nika-cli untouched post-merge (the context lines verified
byte-identical to the CI's ubuntu artifact) · nika-providers +82 lines,
spot-checked against the ubuntu canonical forms (core::time::Duration ·
std::sync::mpsc — no mac skew on the new block).

Gates: crate-size OK (14.9k) · 346+139 lib green · smoke green ·
clippy -D 0 both crates.
