+++ crates/nika-cli/src/verbs/doctor.rs
@@ -471,7 +471,8 @@ mod tests {
     use nika_providers::ProviderRegistry;
 
     use super::*;
-    use crate::verbs::probe::{client_probe_any, env_present, ping_addr, spawn_ping};
+    use crate::verbs::probe::client_probe_any;
+    use nika_providers::probe::{env_present, ping_addr, spawn_ping};
 
     fn cloud(id: &str, var: &str, present: bool) -> ProviderProbe {
         ProviderProbe {
