+++ crates/nika-providers/public-api.txt
@@ -1,4 +1,86 @@
 pub mod nika_providers
+pub mod nika_providers::probe
+pub enum nika_providers::probe::PingState
+pub nika_providers::probe::PingState::Reachable(u64)
+pub nika_providers::probe::PingState::Unreachable
+impl core::clone::Clone for nika_providers::probe::PingState
+pub fn nika_providers::probe::PingState::clone(&self) -> nika_providers::probe::PingState
+impl core::cmp::Eq for nika_providers::probe::PingState
+impl core::cmp::PartialEq for nika_providers::probe::PingState
+pub fn nika_providers::probe::PingState::eq(&self, other: &nika_providers::probe::PingState) -> bool
+impl core::fmt::Debug for nika_providers::probe::PingState
+pub fn nika_providers::probe::PingState::fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result
+impl core::marker::Copy for nika_providers::probe::PingState
+impl core::marker::StructuralPartialEq for nika_providers::probe::PingState
+impl<D> owo_colors::OwoColorize for nika_providers::probe::PingState
+impl<T, U> core::convert::Into<U> for nika_providers::probe::PingState where U: core::convert::From<T>
+pub fn nika_providers::probe::PingState::into(self) -> U
+impl<T, U> core::convert::TryFrom<U> for nika_providers::probe::PingState where U: core::convert::Into<T>
+pub type nika_providers::probe::PingState::Error = core::convert::Infallible
+pub fn nika_providers::probe::PingState::try_from(value: U) -> core::result::Result<T, <T as core::convert::TryFrom<U>>::Error>
+impl<T, U> core::convert::TryInto<U> for nika_providers::probe::PingState where U: core::convert::TryFrom<T>
+pub type nika_providers::probe::PingState::Error = <U as core::convert::TryFrom<T>>::Error
+pub fn nika_providers::probe::PingState::try_into(self) -> core::result::Result<U, <U as core::convert::TryFrom<T>>::Error>
+impl<T> alloc::borrow::ToOwned for nika_providers::probe::PingState where T: core::clone::Clone
+pub type nika_providers::probe::PingState::Owned = T
+pub fn nika_providers::probe::PingState::clone_into(&self, target: &mut T)
+pub fn nika_providers::probe::PingState::to_owned(&self) -> T
+impl<T> core::any::Any for nika_providers::probe::PingState where T: 'static + ?core::marker::Sized
+pub fn nika_providers::probe::PingState::type_id(&self) -> core::any::TypeId
+impl<T> core::borrow::Borrow<T> for nika_providers::probe::PingState where T: ?core::marker::Sized
+pub fn nika_providers::probe::PingState::borrow(&self) -> &T
+impl<T> core::borrow::BorrowMut<T> for nika_providers::probe::PingState where T: ?core::marker::Sized
+pub fn nika_providers::probe::PingState::borrow_mut(&mut self) -> &mut T
+impl<T> core::clone::CloneToUninit for nika_providers::probe::PingState where T: core::clone::Clone
+pub unsafe fn nika_providers::probe::PingState::clone_to_uninit(&self, dest: *mut u8)
+impl<T> core::convert::From<T> for nika_providers::probe::PingState
+pub fn nika_providers::probe::PingState::from(t: T) -> T
+impl<T> nika_error::traits::AsAny for nika_providers::probe::PingState where T: 'static
+pub fn nika_providers::probe::PingState::as_any(&self) -> &(dyn core::any::Any + 'static)
+pub struct nika_providers::probe::ProviderProbe
+pub nika_providers::probe::ProviderProbe::fix_var: alloc::string::String
+pub nika_providers::probe::ProviderProbe::id: alloc::string::String
+pub nika_providers::probe::ProviderProbe::key_present: bool
+pub nika_providers::probe::ProviderProbe::requires_key: bool
+pub nika_providers::probe::ProviderProbe::structured_native: bool
+impl core::clone::Clone for nika_providers::probe::ProviderProbe
+pub fn nika_providers::probe::ProviderProbe::clone(&self) -> nika_providers::probe::ProviderProbe
+impl core::cmp::Eq for nika_providers::probe::ProviderProbe
+impl core::cmp::PartialEq for nika_providers::probe::ProviderProbe
+pub fn nika_providers::probe::ProviderProbe::eq(&self, other: &nika_providers::probe::ProviderProbe) -> bool
+impl core::fmt::Debug for nika_providers::probe::ProviderProbe
+pub fn nika_providers::probe::ProviderProbe::fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result
+impl core::marker::StructuralPartialEq for nika_providers::probe::ProviderProbe
+impl<D> owo_colors::OwoColorize for nika_providers::probe::ProviderProbe
+impl<T, U> core::convert::Into<U> for nika_providers::probe::ProviderProbe where U: core::convert::From<T>
+pub fn nika_providers::probe::ProviderProbe::into(self) -> U
+impl<T, U> core::convert::TryFrom<U> for nika_providers::probe::ProviderProbe where U: core::convert::Into<T>
+pub type nika_providers::probe::ProviderProbe::Error = core::convert::Infallible
+pub fn nika_providers::probe::ProviderProbe::try_from(value: U) -> core::result::Result<T, <T as core::convert::TryFrom<U>>::Error>
+impl<T, U> core::convert::TryInto<U> for nika_providers::probe::ProviderProbe where U: core::convert::TryFrom<T>
+pub type nika_providers::probe::ProviderProbe::Error = <U as core::convert::TryFrom<T>>::Error
+pub fn nika_providers::probe::ProviderProbe::try_into(self) -> core::result::Result<U, <U as core::convert::TryFrom<T>>::Error>
+impl<T> alloc::borrow::ToOwned for nika_providers::probe::ProviderProbe where T: core::clone::Clone
+pub type nika_providers::probe::ProviderProbe::Owned = T
+pub fn nika_providers::probe::ProviderProbe::clone_into(&self, target: &mut T)
+pub fn nika_providers::probe::ProviderProbe::to_owned(&self) -> T
+impl<T> core::any::Any for nika_providers::probe::ProviderProbe where T: 'static + ?core::marker::Sized
+pub fn nika_providers::probe::ProviderProbe::type_id(&self) -> core::any::TypeId
+impl<T> core::borrow::Borrow<T> for nika_providers::probe::ProviderProbe where T: ?core::marker::Sized
+pub fn nika_providers::probe::ProviderProbe::borrow(&self) -> &T
+impl<T> core::borrow::BorrowMut<T> for nika_providers::probe::ProviderProbe where T: ?core::marker::Sized
+pub fn nika_providers::probe::ProviderProbe::borrow_mut(&mut self) -> &mut T
+impl<T> core::clone::CloneToUninit for nika_providers::probe::ProviderProbe where T: core::clone::Clone
+pub unsafe fn nika_providers::probe::ProviderProbe::clone_to_uninit(&self, dest: *mut u8)
+impl<T> core::convert::From<T> for nika_providers::probe::ProviderProbe
+pub fn nika_providers::probe::ProviderProbe::from(t: T) -> T
+impl<T> nika_error::traits::AsAny for nika_providers::probe::ProviderProbe where T: 'static
+pub fn nika_providers::probe::ProviderProbe::as_any(&self) -> &(dyn core::any::Any + 'static)
+pub fn nika_providers::probe::collect_local_pings(registry: &nika_providers::registry::ProviderRegistry, image_url: core::option::Option<&str>, tts_url: core::option::Option<&str>) -> alloc::vec::Vec<(alloc::string::String, alloc::string::String, nika_providers::probe::PingState)>
+pub fn nika_providers::probe::collect_provider_probes(registry: &nika_providers::registry::ProviderRegistry) -> alloc::vec::Vec<nika_providers::probe::ProviderProbe>
+pub fn nika_providers::probe::env_present(name: &str) -> bool
+pub fn nika_providers::probe::ping_addr(url: &str) -> core::option::Option<alloc::string::String>
+pub fn nika_providers::probe::spawn_ping(addr: &str, timeout: core::time::Duration) -> std::sync::mpsc::Receiver<core::option::Option<u64>>
 pub mod nika_providers::profile
 #[non_exhaustive] pub enum nika_providers::profile::WireFormat
 pub nika_providers::profile::WireFormat::Anthropic
